fullstack/backend/requirements.txt
# Core
Flask==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.1
Flask-Cors==4.0.1

# LangChain & ecosystem
langchain==0.2.17
langchain-community==0.2.19
langchain-huggingface==0.0.3
langchain-openai==0.1.25
langchain-core==0.2.43
langchain-chroma==0.1.4

# Embeddings / models
sentence-transformers==3.2.0
transformers==4.44.2
tokenizers==0.19.1
# CPU build of PyTorch (Ubuntu 24.04 + Py311)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.1+cpu ; platform_system == "Linux"
packaging==24.2

# Vector DB
chromadb==0.5.11

# PDF & utils
pypdf==5.1.0
tqdm==4.66.5
beautifulsoup4==4.12.3
lxml==5.2.2

# Testing
pytest==8.3.3
pytest-cov==5.0.0
requests==2.32.3

# Stability pins for 24.04 toolchain (avoid ABI mismatches)
numpy==1.26.4
protobuf==4.25.3
grpcio==1.66.1
pydantic==2.9.1
typing-extensions==4.12.2